requirements.txt
torch>=2.0.0
torchvision>=0.15.0
albumentations>=1.3.0
timm>=0.9.0
opencv-python>=4.7.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
tqdm>=4.65.0
PyYAML>=6.0
